2nd comscie/Safety Incident Reporting App/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.9.2
pydantic-settings==2.6.1
cryptography==43.0.1
python-dotenv==1.0.0
